+++ spec/dummy/app/models/featured_article.rb
@@ -1,0 +1,2 @@
+class FeaturedArticle < Article
+end